refactor: delete proofs and fix tests
