+++ tests/data/credentials.ts
@@ -70,34 +70,6 @@ export const mockVC = {
   },
 };
 
-export const eidasVerifiableCredential = {
-  "@context": [
-    "https://www.w3.org/2018/credentials/v1",
-    "https://www.w3.org/2018/credentials/examples/v1",
-    "https://validatedid.github.io/jsonld-contexts/ades-signatures/v1/",
-  ],
-  id: "http://example.gov/credentials/3732",
-  type: ["VerifiableCredential", "UniversityDegreeCredential"],
-  issuer: "did:key:z6MkmVTtS1RGoxyoGtB6HLMZAEVgYzwnUdD1VkojPj87hAP4",
-  issuanceDate: "2010-01-01T19:24:24Z",
-  credentialSubject: {
-    id: "did:key:z6MknGc3ocHs3zdPiJbnaaqDi58NGb4pk1Sp9WxWufuXSdxf",
-    degree: {
-      type: "BachelorDegree",
-      name: "Bachelor of Science and Arts",
-    },
-  },
-  proof: {
-    type: "CAdESRSASignature2020",
-    created: "2021-03-08T13:09:07Z",
-    proofPurpose: "assertionMethod",
-    verificationMethod:
-      "did:key:z6MkoJagV5bamry562nSX13xhRahYTrnKrujPu6PbjvAccXG#MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAziPjJZebqdAhKZI0zUqd7439PGAlGY/MEIYl/hd0f3paG6UNcnQr74mzkOsJblSTcfzWCPefqWDQ1L/jQ0Iz1gR47I5l+sW7bBRivs1iWON6bcRXuitmVXFJPTl+R6x5vcEcOYlrKq0Fo9fEgbHb5i7Hjl0nWSOLcFsFkgSXfgmv3eI4uYlQyqs6qyICpApcWrijjIRlBlhsbv6Oz0Z8yi0v8FvgEy7Dq40yIoY3IWEGB8GNgi/E3lnLSRsGvzTWT5SXuurSzxrArh2okcNwtAt9DRF3VtMBO701zC2rfI65VyoIKX4QvctGiLSNOSYajDJ9Jg6GjnlE/US0Z6hyDwIDAQAB",
-    cades:
-      "-----BEGIN PKCS7-----MIIKmwYJKoZIhvcNAQcCoIIKjDCCCogCAQExDzANBglghkgBZQMEAgEFADBPBgkqhkiG9w0BBwGgQgRAHQ2KKMMyZK8CJ/2eHacksrQWw6lpnvBKTBypNaLjXJIo4x86yBXNRR33U+VLXMPeHs4eu4YqD5dJqE5E7xsBvqCCB4kwggeFMIIGbaADAgECAhBKco+mAJBghF+KrSdHscajMA0GCSqGSIb3DQEBCwUAMEcxCzAJBgNVBAYTAkVTMREwDwYDVQQKDAhGTk1ULVJDTTElMCMGA1UECwwcQUMgQ29tcG9uZW50ZXMgSW5mb3Jtw6F0aWNvczAeFw0yMDEwMTcwODM2NTVaFw0yMzEwMTcwODM2NTRaMIGgMQswCQYDVQQGEwJFUzEPMA0GA1UEBwwGTUFEUklEMRkwFwYDVQQKDBBGTk1ULVJDTSBQUlVFQkFTMQ4wDAYDVQQLDAVDRVJFUzESMBAGA1UEBRMJUTAwMDAwMDBKMRgwFgYDVQRhDA9WQVRFUy1RMDAwMDAwMEoxJzAlBgNVBAMMHlNFTExPIENPTVBPTkVOVEUgUFJVRUJBUyBFSURBUzCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAM4j4yWXm6nQISmSNM1Kne+N/TxgJRmPzBCGJf4XdH96WhulDXJ0K++Js5DrCW5Uk3H81gj3n6lg0NS/40NCM9YEeOyOZfrFu2wUYr7NYljjem3EV7orZlVxST05fkeseb3BHDmJayqtBaPXxIGx2+Yux45dJ1kji3BbBZIEl34Jr93iOLmJUMqrOqsiAqQKXFq4o4yEZQZYbG7+js9GfMotL/Bb4BMuw6uNMiKGNyFhBgfBjYIvxN5Zy0kbBr801k+Ul7rq0s8awK4dqJHDcLQLfQ0Rd1bTATu9Ncwtq3yOuVcqCCl+EL3LRoi0jTkmGowyfSYOho55RP1EtGeocg8CAwEAAaOCBBEwggQNMAwGA1UdEwEB/wQCMAAwgYEGCCsGAQUFBwEBBHUwczA7BggrBgEFBQcwAYYvaHR0cDovL29jc3Bjb21wLmNlcnQuZm5tdC5lcy9vY3NwL09jc3BSZXNwb25kZXIwNAYIKwYBBQUHMAKGKGh0dHA6Ly93d3cuY2VydC5mbm10LmVzL2NlcnRzL0FDQ09NUC5jcnQwggE0BgNVHSAEggErMIIBJzCCARgGCisGAQQBrGYDCRMwggEIMCkGCCsGAQUFBwIBFh1odHRwOi8vd3d3LmNlcnQuZm5tdC5lcy9kcGNzLzCB2gYIKwYBBQUHAgIwgc0MgcpDZXJ0aWZpY2FkbyBjdWFsaWZpY2FkbyBkZSBzZWxsbyBlbGVjdHLDs25pY28gc2Vnw7puIHJlZ2xhbWVudG8gZXVyb3BlbyBlSURBUy4gU3VqZXRvIGEgbGFzIGNvbmRpY2lvbmVzIGRlIHVzbyBleHB1ZXN0YXMgZW4gbGEgRFBDIGRlIEZOTVQtUkNNIGNvbiBOSUY6IFEyODI2MDA0LUogKEMvSm9yZ2UgSnVhbiAxMDYtMjgwMDktTWFkcmlkLUVzcGHDsWEpMAkGBwQAi+xAAQEwPAYDVR0RBDUwM6QxMC8xLTArBgkrBgEEAaxmAQgMHlNFTExPIENPTVBPTkVOVEUgUFJVRUJBUyBFSURBUzAdBgNVHSUEFjAUBggrBgEFBQcDAgYIKwYBBQUHAwQwDgYDVR0PAQH/BAQDAgXgMB0GA1UdDgQWBBQv7tIRH+uNhu0BIY64efKwhBpWczCBsAYIKwYBBQUHAQMEgaMwgaAwCAYGBACORgEBMAsGBgQAjkYBAwIBDzATBgYEAI5GAQYwCQYHBACORgEGAjByBgYEAI5GAQUwaDAyFixodHRwczovL3d3dy5jZXJ0LmZubXQuZXMvcGRzL1BEU19DT01QX2VzLnBkZhMCZXMwMhYsaHR0cHM6Ly93d3cuY2VydC5mbm10LmVzL3Bkcy9QRFNfQ09NUF9lbi5wZGYTAmVuMB8GA1UdIwQYMBaAFBn4WC8U1qbMmwSYCA1M16sAp4NlMIHgBgNVHR8EgdgwgdUwgdKggc+ggcyGgZ5sZGFwOi8vbGRhcGNvbXAuY2VydC5mbm10LmVzL0NOPUNSTDEsT1U9QUMlMjBDb21wb25lbnRlcyUyMEluZm9ybWF0aWNvcyxPPUZOTVQtUkNNLEM9RVM/Y2VydGlmaWNhdGVSZXZvY2F0aW9uTGlzdDtiaW5hcnk/YmFzZT9vYmplY3RjbGFzcz1jUkxEaXN0cmlidXRpb25Qb2ludIYpaHR0cDovL3d3dy5jZXJ0LmZubXQuZXMvY3Jsc2NvbXAvQ1JMMS5jcmwwDQYJKoZIhvcNAQELBQADggEBAAW4Yb3h18C01LoMFuicbYI9ZxltqfQcz5konRsl6+BvxG8ZJ2nR9bnzEmE2Xl6pziJYI6uz/jOmBtvwLdVFeH2C9Gw1PdcTVZAFPR9FvqCBPzgwO42uoQ5ynmCcnCu/5qRdl+8Seh1ky2Zo88Douv5BwlCZ4V++HIENUfQLRsIX3vh602XsH6iJnwS1tBR/2hucD3yi2ZWzshHsU15aDtYgJ9pLsWjV1KNK2Rdf2k5F9E4Hi2VfJJxz8WaqT6lDmWb18eTrYUPj3NXqIzu4X861+K1om2uu36vkz2QqApr5m8UzTRC5PdfcIslNMGuvF8tHz+ZbPx2y5jtT5XeCJZUxggKSMIICjgIBATBbMEcxCzAJBgNVBAYTAkVTMREwDwYDVQQKDAhGTk1ULVJDTTElMCMGA1UECwwcQUMgQ29tcG9uZW50ZXMgSW5mb3Jtw6F0aWNvcwIQSnKPpgCQYIRfiq0nR7HGozANBglghkgBZQMEAgEFAKCCAQgwGAYJKoZIhvcNAQkDMQsGCSqGSIb3DQEHATAcBgkqhkiG9w0BCQUxDxcNMjEwMzA4MTMwOTA3WjAvBgkqhkiG9w0BCQQxIgQgzDQxqjcBYjeHu4ej6RSyqo8dcBXicF/ywNykGMV5XHQwgZwGCyqGSIb3DQEJEAIvMYGMMIGJMIGGMIGDBCCJL/dZGq3REOHsToCo9jyKvG82/OKynwMZuuWrOnvW9DBfMEukSTBHMQswCQYDVQQGEwJFUzERMA8GA1UECgwIRk5NVC1SQ00xJTAjBgNVBAsMHEFDIENvbXBvbmVudGVzIEluZm9ybcOhdGljb3MCEEpyj6YAkGCEX4qtJ0exxqMwDQYJKoZIhvcNAQELBQAEggEAgqFX2uW123afjmR/CCNNf9+QkIPxm53+GzJc4eFrjscFyn0eZrRvdWSccSpKm3NNoiD8vgSBTIXu8d8RaFJVCJ91VAQwOh3Qp//h82d2IfCBENH15seGR2XW/BqpxHi0BtV8pR9AoCLIIH7YA1zw8aZddaZtTD1y+ZSja+/SM0NpjhedV8NX7byCJOUFTC5N/xKumzG+Kd/TDmerRHbusuP+S3K6ylPlJxl13Q43CjpSTiSNsTGJML8Hcz9Eq7stpUW3fCqBDz53NvIEUsq8asT8bPs/dFkifyW0nTqkRIZ5BgxvLWBPZHZVu9/028DGn1qevCYg1JGL3vselKvQow==-----END PKCS7-----",
-  },
-};
-
 export const offblocksVerifiableCredential = {
   "@context": [
     "https://www.w3.org/2018/credentials/v1",
