+++ tests/unit/ssi.spec.ts
@@ -1,7 +1,3 @@
-import { normalize, Options } from "jsonld";
-import crypto from "crypto";
-import { indication } from "../../src/dtos";
-import { ApiErrorMessages, BadRequestError } from "../../src/errors";
 import {
   calculateLdProofHashforVerification,
   canonizeCredential,
@@ -9,7 +5,6 @@ import {
   getKidFromDidAndPemCertificate,
 } from "../../src/utils/ssi";
 import * as mockedData from "../data/credentials";
-import { getEcontentFromCAdES } from "../../src/libs/secureEnclave/cades";
 
 describe("ssi util tests should", () => {
   it("build a kid", () => {
@@ -46,21 +41,9 @@ describe("ssi util tests should", () => {
       "
     `);
   }, 30000);
-  it("throw BadRequestError when canonize a non Credential", async () => {
-    expect.assertions(1);
-    const expectedError = new BadRequestError(indication.VERIFICATION_FAIL, {
-      detail: ApiErrorMessages.CANONIZE_BAD_PARAMS,
-    });
-    await expect(canonizeCredential({ data: "some data" })).rejects.toThrow(
-      expectedError
-    );
-  });
 
   it("canonize a proofOption", async () => {
-    const proofToNormalize = {
-      ...mockedData.proof,
-    };
-    const canonized = await canonizeProofOptions(proofToNormalize);
+    const canonized = await canonizeProofOptions(mockedData.proof);
     expect(canonized.length).toBeGreaterThan(0);
     expect(canonized)
       .toMatch(`_:c14n0 <http://purl.org/dc/terms/created> "2021-03-05T12:07:47Z"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
@@ -72,11 +55,11 @@ _:c14n0 <https://w3id.org/security#verificationMethod> <did:key:z6Mko3ZSkBCqcFJp
   });
 
   it("calculate LD Proof hash for signature verification", async () => {
-    const credential = mockedData.eidasVerifiableCredential;
+    const credential = mockedData.offblocksVerifiableCredential;
     const { proof } = credential;
 
     const expectedResult =
-      "HQ2KKMMyZK8CJ/2eHacksrQWw6lpnvBKTBypNaLjXJIo4x86yBXNRR33U+VLXMPeHs4eu4YqD5dJqE5E7xsBvg==";
+      "MWVUt3Lz5TdjYAqund/b3FqaSLYDkB6cA08KVM02M7kx6p9Bh8utllQ+b6nH6txElo/yDen7GWKt3eZsFOqVQw==";
 
     const concatenatedHashes = await calculateLdProofHashforVerification(
       credential,
